Ensure workflows not used from processing components
Removed solve_image since it duplicates pipeline capability.
