--- tests/workflows/test_json_helpers.json
@@ -1,98 +1,0 @@
-{
-    "execute": {
-        "use_dask": false,
-        "n_workers": 4,
-        "memory_limit": 4000000000
-    },
-    "component": {
-        "framework": "ARL",
-        "name": "test_wrappers"
-    },
-    "logging": {
-        "filename": "test_wrappers.log",
-        "filemode": "w",
-        "format": "%(asctime)s,%(msecs)d %(name)s %(levelname)s %(message)s",
-        "datefmt": "%H:%M:%S",
-        "level": "INFO"
-    },
-    "buffer": {
-        "directory": "test_results/"
-    },
-    "inputs": {
-        "vis_list": "gleam_perfect_vislist.hdf"
-    },
-    "outputs": {
-        "restored": "gleam_perfect_restored.fits",
-        "residual": "gleam_perfect_residual.fits",
-        "deconvolved": "gleam_perfect_deconvolved.fits"
-    },
-    "imaging": {
-        "context": "wstack",
-        "vis_slices": 11
-    },
-    "image": {
-        "npixel": 1024,
-        "nchan": 1,
-        "polarisation_frame": "stokesI",
-        "cellsize": {
-            "value": 0.001,
-            "unit": "rad"
-        },
-        "phasecentre": {
-            "ra": {
-                "value": 30.0,
-                "unit": "deg"
-            },
-            "dec": {
-                "value": -60.0,
-                "unit": "deg"
-            },
-            "frame": "icrs",
-            "equinox": "j2000"
-        }
-    },
-    "deconvolution": {
-        "scales": [
-            0,
-            3,
-            10
-        ],
-        "algorithm": "mmclean",
-        "psf_support": 64,
-        "nmoment": 3,
-        "niter": 1000,
-        "fractional_threshold": 0.1,
-        "threshold": 0.001,
-        "gain": 0.1,
-        "deconvolve_facets": 8,
-        "deconvolve_overlap": 16,
-        "deconvolve_taper": "tukey",
-        "nmajor": 5
-    },
-    "create_vislist": {
-        "configuration": "LOWBD2",
-        "rmax": 300.0,
-        "frequency": {
-            "start": 0.9e8,
-            "stop": 1.1e8,
-            "steps": 7
-        },
-        "times": {
-            "start": -1.0,
-            "stop": 1.0,
-            "steps": 11
-        },
-        "phasecentre": {
-            "ra": {
-                "value": 30.0,
-                "unit": "deg"
-            },
-            "dec": {
-                "value": -60.0,
-                "unit": "deg"
-            },
-            "frame": "icrs",
-            "equinox": "j2000"
-        }
-    }
-}
